Save coordinator in data.hass
In the next commit, I will add in the switch platform for adding a switch for the "Use spare PV to heat tank" control in the tank. This will also need the coordinator, and we won't want to make two coordinators. So pull out the coordinator made for the sensor platform, and save it in `hass.data`.