Add in toggle switch for solar PV divert option
This adds the first usage of the switch platform in the integration, for adding in a control for the "Use spare PV to heat tank" option in Mixergy tank settings.

I've tested this with my own tank, and checked that toggling in Home Assistant makes the setting toggle on the Mixergy website UI for my tank.

Fixes #32.